Done Studio 1!

+++ src/studio1/Average.java
@@ -8,7 +8,7 @@ public class Average {
         ArgsProcessor ap = new ArgsProcessor(args);
         int n1 = ap.nextInt("Value for n1?");
         int n2 = ap.nextInt("Value for n2?");
-        double average = (n1+n2)/2;
+        double average = (n1+n2)/(double)2;
         System.out.println("Average of " +n1+" and "+n2+" is "+average);
     }
 }
